Upgrade projects to .NET 8 and improve tests
- Updated `FileServiceTests.cs` to include `System.Data.Common` and changed assertions for file existence.
- Enhanced `Dispose` method for better resource management.
- Upgraded target framework for test projects to `net8.0-windows10.0.19041.0` and updated package references.
- Changed `Newtonsoft.Json` version in `GameLibrary.Core.csproj` to `13.0.3`.
- Updated assertions in `PagesTests.cs` for consistency.
- Updated `GameLibrary.csproj` to target `net8.0` and upgraded several package references.

+++ GameLibrary.Core.Tests.xUnit/FileServiceTests.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Data.Common;
+using System.IO;
 
 using GameLibrary.Core.Services;
 
@@ -38,7 +39,7 @@ public class FileServiceTests : IDisposable
         }
         else
         {
-            Assert.True(false, $"File not exist: {_filePath}");
+            Assert.Fail($"File not exist: {_filePath}");
         }
     }
 
@@ -76,9 +77,18 @@ public class FileServiceTests : IDisposable
 
     public void Dispose()
     {
-        if (File.Exists(_filePath))
-        {
-            File.Delete(_filePath);
-        }
-    }
+		Dispose(true);
+		GC.SuppressFinalize(this);
+	}
+
+	protected virtual void Dispose(bool disposing)
+	{
+		if (disposing)
+		{
+			if (File.Exists(_filePath))
+			{
+				File.Delete(_filePath);
+			}
+		}
+	}
 }
